backend/MS-Interview/tools/stress-inference-service/requirements.txt
absl-py==2.4.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
cffi==2.0.0
click==8.3.2
colorama==0.4.6
contourpy==1.3.3
cycler==0.12.1
fastapi==0.135.3
flatbuffers==25.12.19
fonttools==4.62.1
h11==0.16.0
httptools==0.7.1
idna==3.11
kiwisolver==1.5.0
matplotlib==3.10.8
mediapipe==0.10.33
numpy==2.4.4
opencv-contrib-python==4.13.0.92
opencv-python-headless==4.13.0.92
packaging==26.0
pillow==12.2.0
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
pyparsing==3.3.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
python-multipart==0.0.24
PyYAML==6.0.3
six==1.17.0
sounddevice==0.5.5
starlette==1.0.0
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.43.0
watchfiles==1.1.1
websockets==16.0
